+++ NEWS.md
@@ -1,3 +1,8 @@
+# PAMmisc 1.8.1
+
+* Updated `decimateWavFiles` to work with wav files with bit rates different
+than 16 (bugs in `seewave` functions do not pass along bit rate properly)
+
 # PAMmisc 1.8.0
 
 * Reworking of enviro functions to make HYCOM datasets easier to use. Will now